+++ src/utils/clients.js
@@ -8,7 +8,6 @@ const {
 } = process.env
 
 const BitcoinRpcProvider = require('@liquality/bitcoin-rpc-provider')
-const BitcoinNodeWalletProvider = require('@liquality/bitcoin-node-wallet-provider')
 const BitcoinJsWalletProvider = require('@liquality/bitcoin-js-wallet-provider')
 const BitcoinSwapProvider = require('@liquality/bitcoin-swap-provider')
 const BitcoinCollateralProvider = require('@atomicloans/bitcoin-collateral-provider')
