+++ src/utils/contracts.js
@@ -1,4 +1,4 @@
-const web3 = require('./web3')
+const getWeb3 = require('./web3')
 
 const schema = {}
 
@@ -8,6 +8,7 @@ schema.sales = require('../abi/sales')
 schema.erc20 = require('../abi/erc20')
 
 function loadObject (type, address) {
+  const web3 = getWeb3()
   return new web3.eth.Contract(schema[type].abi, address)
 }
