+++ test/loan/lender/deploy/fund.js
@@ -9,7 +9,7 @@ const { fundArbiter, fundAgent, generateSecretHashesArbiter, fundWeb3Address, ge
 const { createCustomFund } = require('../setup/fundSetup')
 const { currencies } = require('../../../../src/utils/fx')
 const { numToBytes32 } = require('../../../../src/utils/finance')
-const web3 = require('../../../../src/utils/web3')
+const web3 = require('web3')
 
 const { toWei } = web3.utils
 
@@ -37,7 +37,7 @@ function deployFund (web3Chain) {
 
       const balanceAfter = await token.methods.balanceOf(process.env[`${principal}_LOAN_FUNDS_ADDRESS`]).call()
       const { lender } = await funds.methods.funds(numToBytes32(fundId)).call()
-      
+
       expect(balanceAfter.toString()).to.equal(BN(balanceBefore).plus(toWei(amount.toString(), currencies[principal].unit)).toString())
       expect(lender).to.equal(agentAddress)
     })
