+++ test/loan/lender/setup/fundSetup.js
@@ -4,7 +4,7 @@ const chaiAsPromised = require('chai-as-promised')
 const web3 = require('web3')
 
 const { getWeb3Address } = require('../../util/web3Helpers')
-const { getAgentAddress, getTestObjects, fundTokens } = require('../../loanCommon')
+const { getTestObjects, fundTokens } = require('../../loanCommon')
 const { numToBytes32 } = require('../../../../src/utils/finance')
 const { currencies } = require('../../../../src/utils/fx')
 const { sleep } = require('../../../../src/utils/async')
@@ -13,7 +13,6 @@ const fundFixtures = require('../fixtures/fundFixtures')
 const { toWei } = web3.utils
 
 chai.should()
-const expect = chai.expect
 
 chai.use(chaiHttp)
 chai.use(chaiAsPromised)
@@ -22,12 +21,9 @@ const server = 'http://localhost:3030/api/loan'
 
 async function createCustomFund (web3Chain, arbiterChain, amount, principal) {
   const currentTime = Math.floor(new Date().getTime() / 1000)
-  const agentPrincipalAddress = await getAgentAddress(server)
   const address = await getWeb3Address(web3Chain)
-  const arbiter = await getWeb3Address(arbiterChain)
   const fundParams = fundFixtures.customFundWithFundExpiryIn100Days(currentTime, principal)
-  const { fundExpiry, liquidationRatio, interest, penalty, fee } = fundParams
-  const [ token, funds ] = await getTestObjects(web3Chain, principal, ['erc20', 'funds'])
+  const [token, funds] = await getTestObjects(web3Chain, principal, ['erc20', 'funds'])
   const unit = currencies[principal].unit
   const amountToDeposit = toWei(amount.toString(), unit)
   await fundTokens(address, amountToDeposit, principal)
@@ -45,7 +41,7 @@ async function createCustomFund (web3Chain, arbiterChain, amount, principal) {
 
 async function depositToFund (web3Chain, amount, principal) {
   const address = await getWeb3Address(web3Chain)
-  const [ token, funds ] = await getTestObjects(web3Chain, principal, ['erc20', 'funds'])
+  const [token, funds] = await getTestObjects(web3Chain, principal, ['erc20', 'funds'])
   const unit = currencies[principal].unit
   const amountToDeposit = toWei(amount.toString(), unit)
   await fundTokens(address, amountToDeposit, principal)
@@ -66,14 +62,14 @@ async function checkFundCreated (fundModelId) {
   let fundId
   while (!created) {
     await sleep(1000)
-    let { body } = await chai.request(server).get(`/funds/${fundModelId}`)
+    const { body } = await chai.request(server).get(`/funds/${fundModelId}`)
     const { status } = body
     if (status === 'CREATED') {
       created = true
       fundId = body.fundId
     }
   }
-  
+
   return fundId
 }
